refactor(net.core): :construction: create AggregateRootBase without property tenant
