+++ packages/CodeDesignPlus.Net.Core/src/CodeDesignPlus.Net.Core.Abstractions/IAggregateRoot.cs
@@ -6,8 +6,8 @@
 /// <remarks>
 /// An aggregate root is an entity that serves as the root of an aggregate, which is a cluster of associated objects that are treated as a single unit.
 /// </remarks>
-/// <seealso cref="IEntity"/>
-public interface IAggregateRoot : IEntity
+/// <seealso cref="IEntityBase"/>
+public interface IAggregateRoot : IEntityBase
 {
     /// <summary>
     /// Gets and clears the domain events associated with the aggregate root.
